app/requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
boto3==1.34.113
botocore==1.34.113
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
greenlet==3.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
loguru==0.7.2
Mako==1.3.0
MarkupSafe==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
protobuf==4.24.2
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.30
starlette==0.27.0
tenacity==8.3.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.24.0.post1
